Update
Change rhythm background
Add PauseScreen prefab
Implement pause screen in rhythm
Draft for result screen

+++ Assets/Scripts/PlayerData.cs
@@ -4,9 +4,9 @@
 
     public static int Exp { get; set; }
 
-    public static int MP { get; set; }
+    public static float MP { get; set; }
 
-    public static int HP { get; set; }
+    public static float HP { get; set; }
 
     public static int MapID { get; set; }
 }
